Functions and methods for image manipulation

Repotomata.create_image — Method

create_image(width, height, parameters)

Create the base image with the given width and height. The parameters will be used to generate the Perlin noise and the random points.

Repotomata.generate_images — Method

generate_images(input)

Create the base image and runs the evolutions. Return a Vector of the created images.

Repotomata.newgeneration — Method

newgeneration(image, rule, background_color)

Produce a new image by applying the rule to the given image.

Repotomata.getpixel — Method

getpixel(image, position)

Return the color of the pixel at position in the image.

Repotomata.setpixel! — Method

setpixel!(image, position, color)

Set the color of the pixel at position in the image.