Repotomata main function

Repotomata.repotomata — Method

repotomata(owner::AbstractString, name::AbstractString; <keyword arguments>)
repotomata(ownername::AbstractString; kwargs...)

Generates a cellular automata animation of the given GitHub repository.

Note

When using the second method, the format must be: "owner/name".

Arguments

epochs::Int=10: the number of generations to compute.

width::Integer=500: the output width (height will be automatically computed with the golden ratio).

output::OutputType=raw: the output type.

output_path::String="": the output path when using OutputType.gif.

rule::RuleType=chromatic: the rule to use.

seed_treshold::Real=0.58: the threshold on the Perlin noise used as seed.

background_color::Colorant=black: the background color.

The input types

Repotomata.OutputType — Type

The different output types

raw: will return a Vector of the created images.

viewer: will open the result in a viewer.

gif: will create a gif file at the given output_path.

See also: repotomata

Repotomata.RuleType — Type

The different rules

life: the "game of life" rule.

languages: the languages rule.

See also: repotomata

The viewer utility function

Repotomata.open_viewer — Method

open_viewer(images)

Display the images in an interactive window.